rust_gungraun adapter delegate to rust_gungraun_json, rust_gungraun_stdout
