+++ lib/bencher_json/src/project/report.rs
@@ -91,6 +91,7 @@ const RUST_CRITERION_INT: i32 = 22;
 const RUST_IAI_INT: i32 = 23;
 const RUST_GUNGRAUN_INT: i32 = 24;
 const RUST_GUNGRAUN_JSON_INT: i32 = 25;
+const RUST_GUNGRAUN_STDOUT_INT: i32 = 26;
 const CPP_INT: i32 = 30;
 const CPP_GOOGLE_INT: i32 = 31;
 const CPP_CATCH2_INT: i32 = 32;
@@ -134,6 +135,7 @@ pub enum Adapter {
     #[serde(alias = "rust_iai_callgrind")]
     RustGungraun = RUST_GUNGRAUN_INT,
     RustGungraunJson = RUST_GUNGRAUN_JSON_INT,
+    RustGungraunStdout = RUST_GUNGRAUN_STDOUT_INT,
     Cpp = CPP_INT,
     CppGoogle = CPP_GOOGLE_INT,
     CppCatch2 = CPP_CATCH2_INT,
@@ -182,6 +184,7 @@ impl Adapter {
             | Self::RustIai
             | Self::RustGungraun
             | Self::RustGungraunJson
+            | Self::RustGungraunStdout
             | Self::CppGoogle
             | Self::CppCatch2
             | Self::GoBench
@@ -210,6 +213,7 @@ impl fmt::Display for Adapter {
             Self::RustIai => write!(f, "rust_iai"),
             Self::RustGungraun => write!(f, "rust_gungraun"),
             Self::RustGungraunJson => write!(f, "rust_gungraun_json"),
+            Self::RustGungraunStdout => write!(f, "rust_gungraun_stdout"),
             Self::Cpp => write!(f, "cpp"),
             Self::CppGoogle => write!(f, "cpp_google"),
             Self::CppCatch2 => write!(f, "cpp_catch2"),
@@ -243,8 +247,8 @@ mod adapter {
         DART_BENCHMARK_HARNESS_INT, DART_INT, GO_BENCH_INT, GO_INT, JAVA_INT, JAVA_JMH_INT,
         JS_BENCHMARK_INT, JS_INT, JS_TIME_INT, JS_VITEST_INT, JSON_INT, MAGIC_INT, PYTHON_ASV_INT,
         PYTHON_INT, PYTHON_PYTEST_INT, RUBY_BENCHMARK_INT, RUBY_INT, RUST_BENCH_INT,
-        RUST_CRITERION_INT, RUST_GUNGRAUN_INT, RUST_GUNGRAUN_JSON_INT, RUST_IAI_INT, RUST_INT,
-        SHELL_HYPERFINE_INT, SHELL_INT,
+        RUST_CRITERION_INT, RUST_GUNGRAUN_INT, RUST_GUNGRAUN_JSON_INT, RUST_GUNGRAUN_STDOUT_INT,
+        RUST_IAI_INT, RUST_INT, SHELL_HYPERFINE_INT, SHELL_INT,
     };
 
     #[derive(Debug, thiserror::Error)]
@@ -271,6 +275,7 @@ mod adapter {
                 Self::RustIai => RUST_IAI_INT.to_sql(out),
                 Self::RustGungraun => RUST_GUNGRAUN_INT.to_sql(out),
                 Self::RustGungraunJson => RUST_GUNGRAUN_JSON_INT.to_sql(out),
+                Self::RustGungraunStdout => RUST_GUNGRAUN_STDOUT_INT.to_sql(out),
                 Self::Cpp => CPP_INT.to_sql(out),
                 Self::CppGoogle => CPP_GOOGLE_INT.to_sql(out),
                 Self::CppCatch2 => CPP_CATCH2_INT.to_sql(out),
@@ -312,6 +317,7 @@ mod adapter {
                 RUST_IAI_INT => Ok(Self::RustIai),
                 RUST_GUNGRAUN_INT => Ok(Self::RustGungraun),
                 RUST_GUNGRAUN_JSON_INT => Ok(Self::RustGungraunJson),
+                RUST_GUNGRAUN_STDOUT_INT => Ok(Self::RustGungraunStdout),
                 CPP_INT => Ok(Self::Cpp),
                 CPP_GOOGLE_INT => Ok(Self::CppGoogle),
                 CPP_CATCH2_INT => Ok(Self::CppCatch2),
